+++ subsys/bluetooth/mesh/adv_ext.c
@@ -174,7 +174,7 @@ static void send_pending_adv(struct k_work *work)
 	}
 
 	/* No more pending buffers */
-	if (IS_ENABLED(CONFIG_BT_MESH_PROXY)) {
+	if (IS_ENABLED(CONFIG_BT_MESH_GATT_SERVER)) {
 		BT_DBG("Proxy Advertising");
 		err = bt_mesh_proxy_adv_start();
 		if (!err) {
